Added REST API and made some basic changes
