+++ api/routes/admin_login.js
@@ -1,0 +1,11 @@
+const express = require('express');
+const con = require("../../database/db");
+const router = express.Router();
+
+
+
+
+router.get('/', function (req, res) {
+    res.render('admin_login.hbs');
+  });
+  module.exports = router;
