+++ api/routes/home.js
@@ -1,0 +1,10 @@
+const express = require('express');
+
+const router = express.Router();
+
+
+router.get('/', function (req, res) {
+    res.render('index.hbs');
+  });
+  
+  module.exports = router;
