+++ package.json
@@ -11,6 +11,7 @@
   "dependencies": {
     "express": "^4.17.1",
     "hbs": "^4.1.2",
+    "morgan": "^1.10.0",
     "mysql": "^2.18.1",
     "mysql2": "^2.2.5"
   }
