+++ views/404.hbs
@@ -51,7 +51,7 @@
             </div>
         </header>
         <section
-            style="background-image: url('images/404error.png');background-repeat: no-repeat;background-size: cover;background-color : white;">
+            style="background-image: url('./img/404error.png');background-repeat: no-repeat;background-size: cover;background-color : white;">
         </section>
         <footer>
             <br>
