+++ views/CSS/feedback.css
@@ -23,8 +23,7 @@ nav{
     
 }
 .active{
-    text-decoration: underline;
-    text-decoration-color: black;
+    text-decoration: none;
     color: black;
 }
