+++ views/CSS/registration.css
@@ -19,9 +19,13 @@ nav {
 .nav-link:hover {
     color: #b8b5ff;
 }
+.navbar-brand{
+    font-weight: 500;
+    font-size: larger;
+    letter-spacing: 2px;
+}
 .active{
-    text-decoration: underline;
-    text-decoration-color: black;
+    text-decoration: none;
     color: black;
 }
 .active:hover{
