+++ views/admin_registration.hbs
@@ -39,7 +39,7 @@
 
             <nav class="navbar navbar-expand-lg ">
                 <div class="container-fluid">
-                    <a class="navbar-brand nav-link" href="/"><i class="fas fa-book-reader fa-2x"></i></a>
+                    <a class="navbar-brand nav-link" href="/"><i class="fas fa-book-reader fa-2x">BMSCE Library</i></a>
                     <button class="navbar-toggler" type="button" data-bs-toggle="collapse"
                         data-bs-target="#navbarSupportedContent" aria-controls="navbarSupportedContent"
                         aria-expanded="false" aria-label="Toggle navigation">
@@ -74,36 +74,6 @@
 
     </header>
    
-        {{!-- <header>
-            <div class="full_nav">
-                <div class="logo">
-                    <img src="images/logo.png" alt="">
-                </div>
-                <div class="navbar">
-                    <nav>
-                        <ul>
-                            <li>
-                                <a href="/"><span class="glyphicon glyphicon-home"></span>HOME</a>
-                            </li>
-                            <li>
-                                <a href="/books"><span class="glyphicon glyphicon-book"></span>BOOKS</a>
-                            </li>
-                            <li>
-                                <a href="/admin_login"><span class="glyphicon glyphicon-user"></span>ADMIN_LOGIN</a>
-                            </li>
-                            <li>
-                                <span class="glyphicon glyphicon-edit"></span><a href="/admin_registration"
-                                    class="active">ADMIN_REGISTRATION</a>
-                            </li>
-                            <li>
-                                <a href="/feedback">FEEDBACK</a>
-                            </li>
-                        </ul>
-                    </nav>
-                </div>
-            </div>
-        </header> --}}
-
 
 
         <section id="registration">
@@ -151,26 +121,6 @@
             </div>
         </div>
     </section>
-        {{!-- <section>
-            <br><br><br><br>
-            <div id="box">
-                <br><br>
-                <h1 style="text-align: center;font-size: 50px;letter-spacing: 2px;">Admin registration Form</h1>
-                <br><br><br>
-                <form name="admin_registration" action="/admin_registration" method="POST">
-                    <div class="login">
-                        <input type="text" name="adminname" placeholder="admin name" required=""><br><br>
-                        <input type="email" name="email" placeholder="email" required=""><br><br>
-                        <input type="password" name="password" placeholder="password" required=""><br><br>
-                        <input type="password" name="password" placeholder="verify password" required=""><br><br>
-                        <button type="submit">SignUp</button><span>{{msg1}}</span>
-                    </div>
-                </form>
-            </div>
-            <a href="#" id="scroll_top">
-                <span class="glyphicon glyphicon-arrow-up"></span>
-            </a>
-        </section> --}}
         {{!-- <footer>
             <br>
             <p style="text-align: center;"><a href="index.html" style="color: white;"><span
