+++ views/student_login.hbs
@@ -41,7 +41,7 @@
 
             <nav class="navbar navbar-expand-lg ">
                 <div class="container-fluid">
-                    <a class="navbar-brand nav-link" href="/"><i class="fas fa-book-reader fa-2x"></i></a>
+                    <a class="navbar-brand nav-link" href="/"><i class="fas fa-book-reader fa-2x"></i>BMSCE Library</a>
                     <button class="navbar-toggler" type="button" data-bs-toggle="collapse"
                         data-bs-target="#navbarSupportedContent" aria-controls="navbarSupportedContent"
                         aria-expanded="false" aria-label="Toggle navigation">
