+++ views/student_registration.hbs
@@ -40,7 +40,7 @@
 
             <nav class="navbar navbar-expand-lg ">
                 <div class="container-fluid">
-                    <a class="navbar-brand nav-link" href="/"><i class="fas fa-book-reader fa-2x"></i></a>
+                    <a class="navbar-brand nav-link" href="/"><i class="fas fa-book-reader fa-2x"></i>BMSCE Library</a>
                     <button class="navbar-toggler" type="button" data-bs-toggle="collapse"
                         data-bs-target="#navbarSupportedContent" aria-controls="navbarSupportedContent"
                         aria-expanded="false" aria-label="Toggle navigation">
@@ -74,37 +74,6 @@
         </div>
 
     </header>
-
-    {{!-- <header>
-        <div class="full_nav">
-            <div class="logo">
-                <img src="images/logo.png" alt="">
-            </div>
-            <div class="navbar">
-                <nav>
-                    <ul>
-                        <li>
-                            <a href="/"><span class="glyphicon glyphicon-home"></span>HOME</a>
-                        </li>
-                        <li>
-                            <a href="/books"><span class="glyphicon glyphicon-book"></span>BOOKS</a>
-                        </li>
-                        <li>
-                            <a href="/student_login"><span class="glyphicon glyphicon-user"></span>USER_LOGIN</a>
-                        </li>
-                        <li>
-                            <span class="glyphicon glyphicon-edit"></span><a href="/student_registration"
-                                class="active">USER_REGISTRATION</a>
-                        </li>
-                        <li>
-                            <a href="/feedback"><span class="glyphicon glyphicon-envelope"></span>FEEDBACK</a>
-                        </li>
-                    </ul>
-                </nav>
-            </div>
-        </div>
-    </header> --}}
-
 
     <section id="registration">
         <div class="container-nav">
@@ -151,26 +120,6 @@
             </div>
         </div>
     </section>
-
-    {{!-- <section>
-        <br><br><br><br>
-        <div id="box">
-            <br><br>
-            <h1 style="text-align: center;font-size: 50px;letter-spacing: 2px;">User registration Form</h1>
-            <br><br><br>
-            <form action="/student_registration" method="POST" id="contact">
-                <div class="login">
-                    <input type="text" name="username" placeholder="username" required=""><br><br>
-                    <input type="email" name="email" placeholder="email" required=""><br><br>
-                    <input type="password" name="password" placeholder="password" required=""><br><br>
-                    <input type="password" name="password" placeholder="verify password" required=""><br><br>
-                    <button type="submit">SignUp</button>
-                </div>
-            </form>
-        </div>
-
-    </section> --}}
-
 
     {{!-- <footer>
         <br>
